Remove legacy Rig pose experiment paths
