+++ src/tests/web/test_modules.py
@@ -102,404 +102,6 @@ def test_rig_pose_presets_use_exact_stable_signatures_and_partial_resolution(
     }
 
 
-def test_builtin_arms_up_pose_is_symmetric_deterministic_and_fail_closed(
-        module_page):
-    result = module_page.evaluate("""async () => {
-      const poses = await import('./js/mesh/weight-rig-procedural-poses.js');
-      const makeRig = includeRight => {
-        const entries = [
-          [0, [0, 0, 0]], [1, [0, -2, 0]], [2, [0, 10, 0]],
-          [10, [-2, 6, 0]], [11, [-3.2, 4, 0]], [12, [-3.7, 3, 0]],
-          ...(includeRight ? [[20, [2, 6, 0]], [21, [3.2, 4, 0]],
-            [22, [3.7, 3, 0]]] : []),
-          [30, [.7, -4, 0]], [31, [.7, -6, 0]],
-        ];
-        const parents = {0: null, 1: 0, 2: 0, 10: 0, 11: 10, 12: 11,
-          30: 1, 31: 30};
-        const children = {0: [1, 2, 10, 30], 1: [31], 2: [], 10: [11],
-          11: [12], 12: [], 30: [31], 31: []};
-        if (includeRight) {
-          Object.assign(parents, {20: 0, 21: 20, 22: 21});
-          Object.assign(children, {0: [1, 2, 10, 20, 30], 20: [21],
-            21: [22], 22: []});
-        }
-        const components = [{componentId: 0, rootId: 0,
-          nodeIds: entries.map(([id]) => id), parentById: parents,
-          childrenById: children,
-          depthById: Object.fromEntries(entries.map(([id]) => [id,
-            id === 0 ? 0 : id === 1 || id === 2 || id === 10 || id === 20
-              ? 1 : id === 30 ? 2 : id === 31 ? 3 : 2]))}];
-        const joints = entries.map(([jointId, restCenter]) => ({
-          jointId, signature: `["source#bone=${jointId}"]`, restCenter,
-          restPivot: restCenter,
-        }));
-        const directions = {
-          10: [-1, -.8, .15], 20: [1, -.8, .15],
-        };
-        return {
-          joints, components, defaultComponents: components,
-          defaultRestPivotByJointId: Object.fromEntries(entries),
-          defaultRestDirectionByJointId: directions,
-          forestEdges: [],
-        };
-      };
-      const rig = makeRig(true);
-      const analysis = poses.analyzeHumanoidRestPose(rig);
-      const nullFrameAnalysis = poses.analyzeHumanoidRestPose(rig,
-        {semanticFrame: null});
-      const permuted = {...rig, joints: [...rig.joints].reverse()};
-      const permutedAnalysis = poses.analyzeHumanoidRestPose(permuted);
-      const transformRig = (source, frame) => {
-        const transform = values => [
-          frame.right[0] * values[0] + frame.up[0] * values[1]
-            + frame.forward[0] * values[2],
-          frame.right[1] * values[0] + frame.up[1] * values[1]
-            + frame.forward[1] * values[2],
-          frame.right[2] * values[0] + frame.up[2] * values[1]
-            + frame.forward[2] * values[2],
-        ];
-        return {
-          ...source,
-          joints: source.joints.map(joint => ({...joint,
-            restCenter: transform(joint.restCenter),
-            restPivot: transform(joint.restPivot),
-          })),
-          defaultRestPivotByJointId: Object.fromEntries(
-            Object.entries(source.defaultRestPivotByJointId)
-              .map(([id, value]) => [id, transform(value)])),
-          defaultRestDirectionByJointId: Object.fromEntries(
-            Object.entries(source.defaultRestDirectionByJointId)
-              .map(([id, value]) => [id, transform(value)])),
-        };
-      };
-      const zUpFrame = {up: [0, 0, 1], right: [1, 0, 0],
-        forward: [0, -1, 0]};
-      const zUpAnalysis = poses.analyzeHumanoidRestPose(
-        transformRig(rig, zUpFrame), {semanticFrame: zUpFrame});
-      const zUpGenerated = poses.generateArmsUpPreset(
-        transformRig(rig, zUpFrame), {
-          semanticFrame: zUpFrame, outwardAngleDegrees: 0,
-        });
-      const facingFrame = {up: [0, 1, 0], right: [-1, 0, 0],
-        forward: [0, 0, -1]};
-      const facingAnalysis = poses.analyzeHumanoidRestPose(
-        transformRig(rig, facingFrame), {semanticFrame: facingFrame});
-      const longArms = makeRig(true);
-      [12, 22].forEach(jointId => {
-        const joint = longArms.joints.find(item => item.jointId === jointId);
-        joint.restCenter[0] = jointId === 12 ? -5.5 : 5.5;
-        joint.restPivot[0] = joint.restCenter[0];
-        longArms.defaultRestPivotByJointId[jointId] = joint.restCenter;
-      });
-      const longArmAnalysis = poses.analyzeHumanoidRestPose(longArms);
-      const decoyRig = makeRig(true);
-      const decoyEntries = [
-        [40, [-2, 6, 0]], [41, [-3.2, 4, 0]], [42, [-3.7, 3, 0]],
-        [50, [2, 6, 0]], [51, [3.2, 4, 0]], [52, [3.7, 3, 0]],
-      ];
-      decoyEntries.forEach(([jointId, restCenter]) => {
-        decoyRig.joints.push({jointId,
-          signature: `["source#bone=${jointId}"]`, restCenter,
-          restPivot: restCenter});
-        decoyRig.defaultRestPivotByJointId[jointId] = restCenter;
-      });
-      Object.assign(decoyRig.defaultRestDirectionByJointId, {
-        40: [-1, -.8, .15], 50: [1, -.8, .15],
-      });
-      const decoyComponent = decoyRig.defaultComponents[0];
-      decoyComponent.nodeIds.push(40, 41, 42, 50, 51, 52);
-      Object.assign(decoyComponent.parentById,
-        {40: 0, 41: 40, 42: 41, 50: 0, 51: 50, 52: 51});
-      Object.assign(decoyComponent.childrenById,
-        {0: [...decoyComponent.childrenById[0], 40, 50],
-          40: [41], 41: [42], 42: [], 50: [51], 51: [52], 52: []});
-      Object.assign(decoyComponent.depthById,
-        {40: 1, 41: 2, 42: 2, 50: 1, 51: 2, 52: 2});
-      const decoyAnalysis = poses.analyzeHumanoidRestPose(decoyRig);
-      const reversed = makeRig(true);
-      Object.assign(reversed.defaultComponents[0].parentById,
-        {10: 11, 11: 12, 12: 0, 20: 21, 21: 22, 22: 0});
-      Object.assign(reversed.defaultComponents[0].childrenById,
-        {0: [1, 2, 12, 22, 30], 10: [], 11: [10], 12: [11],
-          20: [], 21: [20], 22: [21]});
-      const reversedAnalysis = poses.analyzeHumanoidRestPose(reversed);
-      const invalid = makeRig(true);
-      invalid.defaultRestDirectionByJointId[10] = [0, 0, 0];
-      const invalidAnalysis = poses.analyzeHumanoidRestPose(invalid);
-      const unavailable = poses.analyzeHumanoidRestPose(makeRig(false));
-      const descriptor = poses.getBuiltInRigPoseDescriptors(rig)[0];
-      return {
-        available: analysis.available,
-        confidence: analysis.confidence,
-        signatures: analysis.preset?.joints.map(item => item.joint_signature),
-        rotations: analysis.arms && {
-          negativeX: analysis.arms.negativeX.rotation,
-          positiveX: analysis.arms.positiveX.rotation,
-          negativeTarget: analysis.arms.negativeX.targetDirection,
-          positiveTarget: analysis.arms.positiveX.targetDirection,
-        },
-        deterministic: JSON.stringify(analysis.preset)
-          === JSON.stringify(permutedAnalysis.preset),
-        nullFrameSignatures: nullFrameAnalysis.preset?.joints.map(item =>
-          item.joint_signature),
-        zUpSignatures: zUpAnalysis.preset?.joints.map(item =>
-          item.joint_signature),
-        zUpTargets: zUpAnalysis.arms && {
-          negative: zUpAnalysis.arms.negativeX.targetDirection,
-          positive: zUpAnalysis.arms.positiveX.targetDirection,
-        },
-        zUpGeneratedTargets: zUpGenerated.arms && {
-          negative: zUpGenerated.arms.negativeX.targetDirection,
-          positive: zUpGenerated.arms.positiveX.targetDirection,
-        },
-        facingSignatures: facingAnalysis.preset?.joints.map(item =>
-          item.joint_signature),
-        longArmSignatures: longArmAnalysis.preset?.joints.map(item =>
-          item.joint_signature),
-        longArmCounts: longArmAnalysis.diagnostics && {
-          geometry: longArmAnalysis.diagnostics.geometryCandidateCounts,
-          collapsed: longArmAnalysis.diagnostics.candidateCounts,
-        },
-        decoyAvailable: decoyAnalysis.available,
-        decoyReason: decoyAnalysis.reason,
-        reversedReason: reversedAnalysis.reason,
-        invalidReason: invalidAnalysis.reason,
-        missingReason: unavailable.reason,
-        descriptor: {id: descriptor.id, name: descriptor.name,
-          kind: descriptor.kind, available: descriptor.available,
-          semanticUp: descriptor.diagnostics.semanticFrame.up},
-      };
-    }""")
-    assert result["available"] is False
-    assert result["confidence"] == 0
-    assert result["signatures"] is None
-    assert result["deterministic"] is True
-    assert result["nullFrameSignatures"] is None
-    assert result["zUpSignatures"] is None
-    assert result["zUpTargets"] is None
-    assert result["zUpGeneratedTargets"] is None
-    assert result["facingSignatures"] is None
-    assert result["longArmSignatures"] is None
-    assert result["decoyAvailable"] is False
-    assert result["decoyReason"] == "semantic_landmarks_incomplete"
-    assert result["missingReason"] == "semantic_landmarks_incomplete"
-    assert result["reversedReason"] == "semantic_landmarks_incomplete"
-    assert result["invalidReason"] == "semantic_landmarks_incomplete"
-    assert result["descriptor"] == {
-        "id": "builtin:arms-up", "name": "Arms Up", "kind": "builtin",
-        "available": False, "semanticUp": [0, 1, 0],
-    }
-
-
-def test_rig_semantics_detects_bilateral_finger_fans(module_page):
-    result = module_page.evaluate("""async () => {
-      const {analyzeRigSemantics} = await import(
-        './js/mesh/weight-rig-semantics.js');
-      const {analyzeHumanoidRestPose} = await import(
-        './js/mesh/weight-rig-procedural-poses.js');
-      const joints = [];
-      const parentById = {};
-      const childrenById = {};
-      const pivots = {};
-      const add = (jointId, center, parentId = null) => {
-        joints.push({jointId, signature: `[\"semantic#bone=${jointId}\"]`,
-          restCenter: center, restPivot: center});
-        parentById[jointId] = parentId;
-        childrenById[jointId] = [];
-        pivots[jointId] = center;
-        if (parentId !== null) childrenById[parentId].push(jointId);
-      };
-      add(0, [0, 0, 0]);
-      add(1, [0, 6, 0], 0);
-      add(2, [0, 12, 0], 1);
-      add(10, [-2, 7, 0], 0);
-      add(11, [-3.1, 5.7, 0], 10);
-      add(12, [-4.1, 4.5, 0], 11);
-      add(20, [2, 7, 0], 0);
-      add(21, [3.1, 5.7, 0], 20);
-      add(22, [4.1, 4.5, 0], 21);
-      const tips = [
-        [-.65, .9, -.2], [-.8, .5, .1], [-.85, .1, .35],
-        [-.8, -.35, .25], [-.55, -.75, -.1],
-      ];
-      tips.forEach((offset, index) => {
-        const leftFirst = [12, 100 + index * 2];
-        const rightFirst = [22, 200 + index * 2];
-        const left = [-4.1 + offset[0] * .5, 4.5 + offset[1] * .5,
-          offset[2] * .5];
-        const right = [4.1 - offset[0] * .5, 4.5 + offset[1] * .5,
-          offset[2] * .5];
-        add(leftFirst[1], left, leftFirst[0]);
-        add(leftFirst[1] + 1,
-          [-4.1 + offset[0], 4.5 + offset[1], offset[2]], leftFirst[1]);
-        add(rightFirst[1], right, rightFirst[0]);
-        add(rightFirst[1] + 1,
-          [4.1 - offset[0], 4.5 + offset[1], offset[2]], rightFirst[1]);
-      });
-      add(30, [-.8, 1, 0], 0);
-      add(31, [.8, 1, 0], 0);
-      const component = {componentId: 0, rootId: 0,
-        nodeIds: joints.map(joint => joint.jointId), parentById,
-        childrenById,
-        depthById: Object.fromEntries(joints.map(joint => [joint.jointId, 1]))};
-      component.depthById[0] = 0;
-      const rig = {joints, components: [component],
-        defaultComponents: [component],
-        defaultRestPivotByJointId: pivots};
-      const transformRig = (source, frame, scale = 1) => {
-        const transform = values => [
-          (frame.right[0] * values[0] + frame.up[0] * values[1]
-            + frame.forward[0] * values[2]) * scale,
-          (frame.right[1] * values[0] + frame.up[1] * values[1]
-            + frame.forward[1] * values[2]) * scale,
-          (frame.right[2] * values[0] + frame.up[2] * values[1]
-            + frame.forward[2] * values[2]) * scale,
-        ];
-        return {
-          ...source,
-          joints: source.joints.map(joint => ({...joint,
-            restCenter: transform(joint.restCenter),
-            restPivot: transform(joint.restPivot),
-          })),
-          defaultRestPivotByJointId: Object.fromEntries(
-            Object.entries(source.defaultRestPivotByJointId)
-              .map(([id, value]) => [id, transform(value)])),
-        };
-      };
-      const zUpFrame = {up: [0, 0, 1], right: [1, 0, 0],
-        forward: [0, -1, 0]};
-      const small = analyzeRigSemantics(transformRig(rig,
-        {up: [0, 1, 0], right: [1, 0, 0], forward: [0, 0, 1]}, .1));
-      const large = analyzeRigSemantics(transformRig(rig,
-        {up: [0, 1, 0], right: [1, 0, 0], forward: [0, 0, 1]}, 10));
-      const brokenChildren = Object.fromEntries(Object.entries(childrenById)
-        .map(([id, children]) => [id, [...children]]));
-      const brokenParentById = {...parentById};
-      delete brokenParentById[11];
-      brokenChildren[10] = [];
-      const brokenComponent = {...component, parentById: brokenParentById,
-        childrenById: brokenChildren};
-      const brokenRig = {...rig, components: [brokenComponent],
-        defaultComponents: [brokenComponent]};
-      const wingJoints = rig.joints.map(joint => ({...joint,
-        restCenter: [...joint.restCenter], restPivot: [...joint.restPivot]}));
-      const wingParentById = {...parentById};
-      const wingChildrenById = Object.fromEntries(Object.entries(childrenById)
-        .map(([id, children]) => [id, [...children]]));
-      const wingPivots = {...pivots};
-      const addWing = (jointId, center, parentId) => {
-        wingJoints.push({jointId, signature: `[\"wing#bone=${jointId}\"]`,
-          restCenter: center, restPivot: center});
-        wingParentById[jointId] = parentId;
-        wingChildrenById[jointId] = [];
-        wingChildrenById[parentId].push(jointId);
-        wingPivots[jointId] = center;
-      };
-      [-1, 1].forEach(side => {
-        const palmId = side < 0 ? 400 : 500;
-        const palm = [side * 7, 6, 8];
-        addWing(palmId, palm, 0);
-        [[-.8, .8], [-.95, .4], [-1.05, 0], [-.95, -.4],
-          [-.8, -.8]].forEach(([lateral, vertical], index) => {
-          const firstId = palmId + 1 + index * 2;
-          addWing(firstId, [side * 7 + side * lateral * .5,
-            6 + vertical * .5, 8], palmId);
-          addWing(firstId + 1, [side * 7 + side * lateral,
-            6 + vertical, 8], firstId);
-        });
-      });
-      const wingComponent = {...component,
-        nodeIds: wingJoints.map(joint => joint.jointId),
-        parentById: wingParentById, childrenById: wingChildrenById};
-      const wingRig = {...rig, joints: wingJoints,
-        components: [wingComponent], defaultComponents: [wingComponent],
-        defaultRestPivotByJointId: wingPivots};
-      const helperJoints = rig.joints.map(joint => ({...joint,
-        restCenter: [...joint.restCenter], restPivot: [...joint.restPivot]}));
-      const helperParentById = {...parentById};
-      const helperChildrenById = Object.fromEntries(Object.entries(childrenById)
-        .map(([id, children]) => [id, [...children]]));
-      const helperPivots = {...pivots};
-      const addHelper = (jointId, center, parentId) => {
-        helperJoints.push({jointId, signature: `["helper#bone=${jointId}"]`,
-          restCenter: center, restPivot: center});
-        helperParentById[jointId] = parentId;
-        helperChildrenById[jointId] = [];
-        helperChildrenById[parentId].push(jointId);
-        helperPivots[jointId] = center;
-      };
-      [-1, 1].forEach(side => {
-        const palmId = side < 0 ? 600 : 700;
-        addHelper(palmId, [side * 3.2, 5.2, 0], 0);
-        [[-.18, .18], [-.2, .06], [-.2, -.06], [-.16, -.18]]
-          .forEach(([lateral, vertical], index) => {
-            const branchId = palmId + 1 + index;
-            addHelper(branchId, [side * 3.2 + side * lateral,
-              5.2 + vertical, 0], palmId);
-          });
-      });
-      const helperComponent = {...component,
-        nodeIds: helperJoints.map(joint => joint.jointId),
-        parentById: helperParentById, childrenById: helperChildrenById};
-      const helperRig = {...rig, joints: helperJoints,
-        components: [helperComponent], defaultComponents: [helperComponent],
-        defaultRestPivotByJointId: helperPivots};
-      return {
-        semantic: analyzeRigSemantics(rig),
-        pose: analyzeHumanoidRestPose(rig),
-        zUp: analyzeRigSemantics(transformRig(rig, zUpFrame),
-          {semanticFrame: zUpFrame}),
-        small,
-        large,
-        broken: {
-          semantic: analyzeRigSemantics(brokenRig),
-          pose: analyzeHumanoidRestPose(brokenRig),
-        },
-        wings: analyzeRigSemantics(wingRig),
-        helpers: analyzeRigSemantics(helperRig),
-      };
-    }""")
-    semantic = result["semantic"]
-    assert semantic["available"], result
-    assert semantic["landmarks"]["negativeHand"]["fingerCount"] >= 4
-    assert semantic["landmarks"]["positiveHand"]["fingerCount"] >= 4
-    assert semantic["landmarks"]["negativeArm"]["shoulder"]["jointId"] == 10
-    assert semantic["landmarks"]["positiveArm"]["shoulder"]["jointId"] == 20
-    assert semantic["landmarks"]["negativeArm"]["hand"]["jointId"] == 12
-    assert semantic["landmarks"]["positiveArm"]["hand"]["jointId"] == 22
-    assert semantic["landmarks"]["negativeArm"]["semanticPathIds"] == [10, 11, 12]
-    assert semantic["landmarks"]["positiveArm"]["semanticPathIds"] == [20, 21, 22]
-    assert result["pose"]["available"], result["pose"]
-    assert result["pose"]["diagnostics"]["detector"] == "semantic-hands"
-    assert [item["joint_signature"]
-            for item in result["pose"]["preset"]["joints"]] == [
-        '["semantic#bone=10"]', '["semantic#bone=20"]']
-    assert result["pose"]["arms"]["negativeX"]["targetDirection"][1] > .9
-    assert result["pose"]["arms"]["positiveX"]["targetDirection"][1] > .9
-    assert result["pose"]["diagnostics"]["poseValidation"]["negativeX"]["heightGain"] > .08
-    assert result["pose"]["diagnostics"]["poseValidation"]["positiveX"]["sidePreserved"]
-    assert result["zUp"]["available"]
-    assert result["zUp"]["landmarks"]["negativeHand"]["jointId"] == 12
-    assert result["zUp"]["landmarks"]["positiveHand"]["jointId"] == 22
-    assert result["small"]["available"] and result["large"]["available"]
-    assert result["small"]["landmarks"]["negativeHand"]["jointId"] == 12
-    assert result["large"]["landmarks"]["positiveHand"]["jointId"] == 22
-    assert result["broken"]["semantic"]["available"]
-    assert not result["broken"]["semantic"]["landmarks"]["negativeArm"][
-        "poseConnectivity"]["connected"]
-    assert "negative_arm_pose_disconnected" in result["broken"]["semantic"]["issues"]
-    assert not result["broken"]["pose"]["available"]
-    assert result["broken"]["pose"]["reason"] == "arm_pose_connectivity_insufficient"
-    assert result["wings"]["available"]
-    assert result["wings"]["landmarks"]["negativeHand"]["jointId"] == 12
-    assert result["wings"]["landmarks"]["positiveHand"]["jointId"] == 22
-    assert result["helpers"]["available"]
-    assert result["helpers"]["landmarks"]["negativeHand"]["jointId"] == 12
-    assert result["helpers"]["landmarks"]["positiveHand"]["jointId"] == 22
-    assert result["helpers"]["landmarks"]["negativeArm"]["shoulder"]["jointId"] == 10
-    assert result["helpers"]["landmarks"]["positiveArm"]["shoulder"]["jointId"] == 20
-
-
 def test_rig_overlay_reuses_forest_buffers_and_model_frame(module_page):
     page = module_page
     result = page.evaluate("""async () => {
