+++ src/web/js/main.js
@@ -31,7 +31,7 @@ import {
   selectRigJoint, setRigJointRotation, setRigJointRoot,
   getRigRotationSnapDegrees, setRigRotationSnapDegrees, setRigVisible,
   setRigOverlayScope,
-  setRigPoseControlStatus, applySavedRigPosePreset, deleteRigPosePreset,
+  setRigPoseControlStatus, deleteRigPosePreset,
   applyRigPosePresetById,
   getRigPresetState, renameRigPosePreset, saveRigPosePreset,
   selectRigPosePreset,
@@ -421,7 +421,6 @@ rendererReady.then(ready => {
     resetRigJoint: {value: resetRigJoint},
     resetRigPose: {value: resetRigPose},
     getRigPresetState: {value: getRigPresetState},
-    applySavedRigPosePreset: {value: applySavedRigPosePreset},
     applyRigPosePresetById: {value: applyRigPosePresetById},
     saveRigPosePreset: {value: saveRigPosePreset},
     renameRigPosePreset: {value: renameRigPosePreset},
